backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
librosa==0.10.1
numpy==1.26.2
replicate==0.20.0
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.2
soundfile==0.12.1
moviepy==1.0.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4 